ACTIN-2000: Remove provided datamodel and schema from actin
Move extraction enums to extractors.
