+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/feed/emc/JacksonSerializable.kt
@@ -1,0 +1,3 @@
+package com.hartwig.actin.clinical.feed.emc
+
+annotation class JacksonSerializable
