+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/feed/emc/digitalfile/DigitalFileEntry.kt
@@ -2,9 +2,9 @@ package com.hartwig.actin.clinical.feed.emc.digitalfile
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
-import com.hartwig.actin.datamodel.clinical.provided.JacksonSerializable
 import com.hartwig.actin.clinical.feed.emc.FeedEntry
 import com.hartwig.actin.clinical.feed.emc.FeedSubjectDeserializer
+import com.hartwig.actin.clinical.feed.emc.JacksonSerializable
 import java.time.LocalDate
 
 @JacksonSerializable
