+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/feed/emc/medication/MedicationEntry.kt
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.DeserializationContext
 import com.fasterxml.jackson.databind.JsonDeserializer
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
-import com.hartwig.actin.datamodel.clinical.provided.JacksonSerializable
 import com.hartwig.actin.clinical.feed.emc.EuropeanDecimalDeserializer
 import com.hartwig.actin.clinical.feed.emc.FeedEntry
 import com.hartwig.actin.clinical.feed.emc.FeedSubjectDeserializer
+import com.hartwig.actin.clinical.feed.emc.JacksonSerializable
 import java.time.LocalDate
 
 class ActiveDeserializer : JsonDeserializer<Boolean>() {
