+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/feed/emc/questionnaire/QuestionnaireEntry.kt
@@ -2,11 +2,11 @@ package com.hartwig.actin.clinical.feed.emc.questionnaire
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
-import com.hartwig.actin.datamodel.clinical.provided.JacksonSerializable
 import com.hartwig.actin.clinical.feed.emc.FeedEntry
 import com.hartwig.actin.clinical.feed.emc.FeedSubjectDeserializer
 import com.hartwig.actin.clinical.feed.emc.FeedValidation
 import com.hartwig.actin.clinical.feed.emc.FeedValidator
+import com.hartwig.actin.clinical.feed.emc.JacksonSerializable
 import java.time.LocalDate
 
 @JacksonSerializable
