+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/feed/standard/extraction/StandardBloodTransfusionExtractor.kt
@@ -2,9 +2,21 @@ package com.hartwig.actin.clinical.feed.standard.extraction
 
 import com.hartwig.actin.clinical.ExtractionResult
 import com.hartwig.actin.clinical.curation.extraction.CurationExtractionEvaluation
-import com.hartwig.actin.datamodel.clinical.provided.ProvidedBloodTransfusionProduct
 import com.hartwig.actin.datamodel.clinical.BloodTransfusion
 import com.hartwig.feed.datamodel.FeedPatientRecord
+
+private enum class BloodTransfusionProduct {
+    PLASMA_A,
+    PLASMA_B,
+    PLASMA_O,
+    PLASMA_AB,
+    PLATELETS_POOLED,
+    PLATELETS_POOLED_RADIATED,
+    ERYTHROCYTES_RADIATED,
+    APHERESIS_PLASMA,
+    ERTHROCYTES_FILTERED,
+    PLATELETS_APHERESIS
+}
 
 class StandardBloodTransfusionExtractor : StandardDataExtractor<List<BloodTransfusion>> {
     override fun extract(ehrPatientRecord: FeedPatientRecord): ExtractionResult<List<BloodTransfusion>> {
@@ -17,11 +29,11 @@ class StandardBloodTransfusionExtractor : StandardDataExtractor<List<BloodTransf
     }
 
     private fun mapTransfusionProduct(input: String): String {
-        val product = enumeratedInput<ProvidedBloodTransfusionProduct>(input)
+        val product = enumeratedInput<BloodTransfusionProduct>(input)
         return when (product) {
-            ProvidedBloodTransfusionProduct.PLASMA_A, ProvidedBloodTransfusionProduct.PLASMA_B, ProvidedBloodTransfusionProduct.PLASMA_O, ProvidedBloodTransfusionProduct.PLASMA_AB, ProvidedBloodTransfusionProduct.APHERESIS_PLASMA -> "Plasma"
-            ProvidedBloodTransfusionProduct.PLATELETS_POOLED, ProvidedBloodTransfusionProduct.PLATELETS_APHERESIS, ProvidedBloodTransfusionProduct.PLATELETS_POOLED_RADIATED -> "Trombocyte"
-            ProvidedBloodTransfusionProduct.ERTHROCYTES_FILTERED, ProvidedBloodTransfusionProduct.ERYTHROCYTES_RADIATED -> "Erythrocytes"
+            BloodTransfusionProduct.PLASMA_A, BloodTransfusionProduct.PLASMA_B, BloodTransfusionProduct.PLASMA_O, BloodTransfusionProduct.PLASMA_AB, BloodTransfusionProduct.APHERESIS_PLASMA -> "Plasma"
+            BloodTransfusionProduct.PLATELETS_POOLED, BloodTransfusionProduct.PLATELETS_APHERESIS, BloodTransfusionProduct.PLATELETS_POOLED_RADIATED -> "Trombocyte"
+            BloodTransfusionProduct.ERTHROCYTES_FILTERED, BloodTransfusionProduct.ERYTHROCYTES_RADIATED -> "Erythrocytes"
             else -> input
         }
     }
