+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/feed/standard/extraction/StandardLabValuesExtractor.kt
@@ -6,7 +6,6 @@ import com.hartwig.actin.clinical.curation.CurationDatabase
 import com.hartwig.actin.clinical.curation.CurationResponse
 import com.hartwig.actin.clinical.curation.config.LabMeasurementConfig
 import com.hartwig.actin.clinical.curation.extraction.CurationExtractionEvaluation
-import com.hartwig.actin.datamodel.clinical.provided.ProvidedLabUnit
 import com.hartwig.actin.datamodel.clinical.LabMeasurement
 import com.hartwig.actin.datamodel.clinical.LabUnit
 import com.hartwig.actin.datamodel.clinical.LabValue
@@ -55,6 +54,64 @@ class StandardLabValuesExtractor(private val labMeasurementCuration: CurationDat
             ProvidedLabUnit.TRILLIONS_PER_LITER -> LabUnit.TRILLIONS_PER_LITER
             ProvidedLabUnit.OTHER -> LabUnit.NONE
             else -> LabUnit.valueOf(ProvidedLabUnit.fromString(it.unit).name)
+        }
+    }
+}
+
+private enum class ProvidedLabUnit(vararg val externalFormats: String) {
+    NANOGRAMS_PER_LITER("ng/L"),
+    NANOGRAMS_PER_MILLILITER("ng/mL"),
+    MICROGRAMS_PER_LITER("ug/L"),
+    MICROGRAMS_PER_MICROLITER("µg/µL"),
+    MILLIGRAMS_PER_DECILITER("mg/dL"),
+    MILLIGRAMS_PER_MILLIMOLE("mg/mmol"),
+    MILLIGRAMS_PER_LITER("mg/L"),
+    GRAMS_PER_DECILITER("g/dL"),
+    GRAMS_PER_LITER("g/L"),
+    GRAMS_PER_MOLE("g/mol"),
+    KILOGRAMS_PER_LITER("kg/L"),
+    MICROGRAMS_PER_GRAM("µg/g"),
+    GRAMS("g"),
+    PICOMOLES_PER_LITER("pmol/L"),
+    NANOMOLES_PER_LITER("nmol/L"),
+    MICROMOLES_PER_LITER("umol/L"),
+    MILLIMOLES_PER_LITER("mmol/L"),
+    MILLIMOLES_PER_MOLE("mmol/mol"),
+    CELLS_PER_CUBIC_MILLIMETER("cells/mm3"),
+    MILLIONS_PER_LITER("10E6/L"),
+    MILLIONS_PER_MILLILITER("10E6/mL"),
+    BILLIONS_PER_LITER("10E9/L"),
+    TRILLIONS_PER_LITER("10E12/L"),
+    MILLIUNITS_PER_LITER("mU/L"),
+    UNITS_PER_LITER("U/L"),
+    UNITS_PER_MILLILITER("U/mL"),
+    KILOUNITS_PER_LITER("kU/L"),
+    INTERNATIONAL_UNITS_PER_LITER("IU/L"),
+    UNITS_OF_INR("INR"),
+    NANOMOLES_PER_DAY("nmol/24h"),
+    MILLIMOLES_PER_DAY("mmol/24h"),
+    MILLIMETERS_PER_HOUR("mm/hr"),
+    MILLILITERS_PER_MINUTE("mL/min"),
+    FEMTOLITERS("fL"),
+    MILLILITERS("mL"),
+    KILO_PASCAL("kPa"),
+    SECONDS("sec"),
+    PERCENTAGE("%"),
+    PERCENTAGE_OF_LEUKOCYTES("% of leukocytes"),
+    PERCENTAGE_OF_T_CELLS("% of T-cells"),
+    MILLI_OSMOLE_PER_KILOGRAM("mOsm/kg"),
+    INTERNATIONAL_UNITS_PER_MILLILITER("IU/ml"),
+    PRNT50("PRNT50"),
+    OTHER,
+    NONE("");
+
+    companion object {
+        fun fromString(input: String?): ProvidedLabUnit {
+            return input?.let { inputString ->
+                ProvidedLabUnit.entries.firstOrNull {
+                    it.externalFormats.map { f -> f.lowercase() }.contains(inputString.lowercase())
+                } ?: OTHER
+            } ?: NONE
         }
     }
 }
