+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/feed/standard/extraction/StandardPatientDetailsExtractor.kt
@@ -4,8 +4,13 @@ import com.hartwig.actin.clinical.ExtractionResult
 import com.hartwig.actin.clinical.curation.extraction.CurationExtractionEvaluation
 import com.hartwig.actin.datamodel.clinical.Gender
 import com.hartwig.actin.datamodel.clinical.PatientDetails
-import com.hartwig.actin.datamodel.clinical.provided.ProvidedGender
 import com.hartwig.feed.datamodel.FeedPatientRecord
+
+private enum class ProvidedGender {
+    MALE,
+    FEMALE,
+    OTHER
+}
 
 class StandardPatientDetailsExtractor : StandardDataExtractor<PatientDetails> {
     override fun extract(ehrPatientRecord: FeedPatientRecord): ExtractionResult<PatientDetails> {
