--- clinical/src/main/resources/json_schema/README.md
@@ -1,8 +1,0 @@
-*** Provided Clinical Data JSON Schema
-
-Contained in this folder is a JSON schema that describes the structure of the clinical data that is expected to be provided to the system. 
-
-To regenerate the schema after a change to any classes in `src/main/kotlin/com/hartwig/actin/clinical/feed/standard/ProvidedDataFeedDatamodel.kt`,
-run `ProvidedClinicalSchemaWriterKt` from IntelliJ. It requires no arguments, so just right click and run. 
-
-When complete, format and commit the schema stored in this directory.
