--- datamodel/src/main/kotlin/com/hartwig/actin/datamodel/clinical/provided/JacksonSerializable.kt
@@ -1,3 +1,0 @@
-package com.hartwig.actin.datamodel.clinical.provided
-
-annotation class JacksonSerializable
